management: add update alert rule APIs
Add PATCH /api/v1/alerting/rules for bulk
update of platform and user-defined alert
rules with drop/restore, label overrides,
and per-rule update support.

Refactor: introduce UpdateAlertRuleLabels
unified method that routes internally to
platform (ARC) or user-defined (PR mutation)
paths, replacing the error-sniffing fallback
pattern in the HTTP handler.

Extend drop/restore to user-defined rules
via ARC. Rename DropPlatformAlertRule and
RestorePlatformAlertRule to DropAlertRule
and RestoreAlertRule.

Add validateDropRestorePreconditions that
checks the PrometheusRule and AlertingRule
CR for GitOps management while still
allowing drops on operator-managed rules
(their whole purpose).

Reject requests that combine
alertingRuleEnabled with labels or
classification — these are mutually
exclusive operations to prevent
partial-apply inconsistencies.

Fix user label updates to read source
labels from the PrometheusRule directly
instead of the relabeled cache, preventing
ARC overlay values from being baked into
the PR source.

Fixes:
- Propagate errors from
  cleanupARCForDeletedRule instead of
  swallowing them (nilerr lint)
- Preserve rule Drops when clearing the
  last label override (prevent silent
  restore)
- Validate ARC ownership on fallback
  restore path (block GitOps-managed ARC
  modification)
- Return errors from findARCByAlertRuleID
  instead of silently continuing
- Use t.Fatalf in get_rule_by_id_test to
  prevent nil deref on type assertion
  failure
- Use httptest.NewRequestWithContext
  (noctx)
- Block user-rule edits when parent
  PrometheusRule is GitOps-managed
- Allow ARC fallback for operator-managed
  rules (the ARC is a separate resource
  not reconciled by the operator)

+++ internal/managementrouter/user_defined_alert_rule_bulk_delete.go
@@ -32,7 +32,7 @@ func (hr *httpRouter) BulkDeleteUserDefinedAlertRules(w http.ResponseWriter, req
 			msg := "missing ruleId"
 			results = append(results, DeleteAlertRuleResult{
 				Id:         id,
-				StatusCode: http.StatusBadRequest,
+				StatusCode: int32(http.StatusBadRequest),
 				Message:    &msg,
 			})
 			continue
@@ -42,14 +42,14 @@ func (hr *httpRouter) BulkDeleteUserDefinedAlertRules(w http.ResponseWriter, req
 			status, message := parseError(err)
 			results = append(results, DeleteAlertRuleResult{
 				Id:         id,
-				StatusCode: status,
+				StatusCode: int32(status),
 				Message:    &message,
 			})
 			continue
 		}
 		results = append(results, DeleteAlertRuleResult{
 			Id:         id,
-			StatusCode: http.StatusNoContent,
+			StatusCode: int32(http.StatusNoContent),
 		})
 	}
